update for windows to make it working with pytesseract and modularization of the code
